refactor: UserModel
